requirements.txt
ansi2html==1.8.0
cachelib==0.9.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dash==2.13.0
dash-core-components==2.0.0
dash-extensions==1.0.3
dash-html-components==2.0.0
dash-table==5.0.0
dataclass-wizard==0.22.2
EditorConfig==0.12.3
Flask==2.2.5
Flask-Caching==2.0.2
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsbeautifier==1.14.9
MarkupSafe==2.1.3
more-itertools==9.1.0
nest-asyncio==1.5.7
numpy==1.26.0
packaging==23.1
pandas==2.1.0
patsy==0.5.3
plotly==5.17.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
retrying==1.3.4
scipy==1.11.2
six==1.16.0
SQLAlchemy==2.0.20
statsmodels==0.14.0
tenacity==8.2.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.2.3
